requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
loguru
chromadb
langchain
pypdf
pandas
pyyaml
rich
pydantic-settings
langchain==0.2.16
langchain-openai==0.1.23
langchain-community
